Update the component name
